HHH-15473 Instantiate collections in the default fetch group by default
