+++ hibernate-core/src/main/java/org/hibernate/boot/SessionFactoryBuilder.java
@@ -638,7 +638,8 @@ public interface SessionFactoryBuilder {
 	 * Should collections be included in the default fetch group when bytecode
 	 * enhancement is used?
 	 *
-	 * @param enabled {@code true} collections should be included
+	 * @param enabled {@code true} collections should be included, {@code false} they should not.
+	 * Default is {@code true}.
 	 *
 	 * @return {@code this}, for method chaining
 	 */
